[FIX] 기록 상세보기 interface 수정 #48
