--- src/interface/record/RecordDatailResponseDto.ts
@@ -1,9 +1,0 @@
-import { CommentDto } from '../comment/CommentDto';
-import { RecordDto } from './RecordDto';
-
-export interface RecordResponseDto {
-  leftId: number;
-  rightId: number;
-  record: RecordDto;
-  comments: CommentDto[];
-}
