+++ src/interface/record/RecordPreviewResponseDto.ts
@@ -1,0 +1,6 @@
+import { RecordDto } from './RecordDto';
+
+export interface RecordPreviewResponseDto {
+  record: RecordDto;
+  commentWriterPhotos: string[];
+}
