+++ src/interface/record/RecordResponseDto.ts
@@ -1,6 +1,9 @@
+import { CommentDto } from '../comment/CommentDto';
 import { RecordDto } from './RecordDto';
 
-export interface RecordPreviewResponseDto {
+export interface RecordResponseDto {
+  leftId: number | null;
+  rightId: number | null;
   record: RecordDto;
-  commentWriterPhotos: string[];
+  comments: CommentDto[];
 }
